FIXUP heap: rename top/bottom to org/end
`org` is the standard terminology we'll use for the unmoving end of memory
allocation areas because this might be the top or the bottom of the memory
range, depending on direction of growth, and will always be the "bottom" of
a stack, even when growing down.

For heaps, the other end of the allocated area is the `brk` ("break") named
after the old Unix brk()/sbrk() calls. This may move in either direction
and there may be unallocated space between the `org` and `brk`.

For stacks, the top of the stack is the `ptr` ("pointer"), regardless of
which direction it grows.

These always point to the _boundary_ between the memory allocation area and
other memory. The higher address points the address of the first-allocated
cell for an area that grows upward and to the address _after_ the first
allocated cell for an area that grows downward. The lower address points to
`org` when nothing has been allocated, the next allocation when growing
upward, and the most recent allocation when growing downward. Thus, it
always holds that the current allocated size is given by `org + brk/ptr`
when growing upward and `org - brk/ptr` when growing downward.