b8tool/bin/b8tool: Simple rebuild skip for asl and asltest
For just asl() and asltest() we check the modification times of the input
and output `.p` files, and skip calling runasl() if the output file is
newer.

This saves only about 25% on overall wall-clock time when doing a build
with no changes: on a set of 30 files on my desktop build time drops from
6.0 to 4.4 seconds. However, this tells us also that our biggest slowdown
is probably the overhead of running b8tool separately for every source
file, so we now know where to go next for a much bigger improvement. (And
of course this current improvement will be a much bigger fraction once we
stop calling b8tool separately.)