b8tool: Add `aslauto` command to discover and build source files
`aslauto` finds and reads .pt files and from their `test_rig` and
`object_files` global properties determines what object files they need and
from what source they should be built, and does the builds. This replaces
many individual `asltest` and `asl` commands. There are a few additional
object files not discovered this way; their source files are all now passed
to a single invocation of the `asl` command which has been changed to
accept a list of source files (building a separate object file for each
one). This has been checked to confirm that it produces the same set of
output files as the previous code that did individual builds.

This speeds up the builds massively as most of the build time was actually
invocation and startup of the Python interpreter and the b8tool modules.

This does not yet cover any ASxxxx builds (we use --exclude to leave them
out of the `aslauto` discovery and still do separate `asx` and `asxlink`
commands for each one). And of course the `buildtoolset` commands are also
still separate. These account for the vast majority of the remaining
overhead.

Testing with Python 3.9 (Python 3.5 is significantly slower) and ASxxxx
builds excluded, and comparing to the original version that always did a
full rebuild (commit 213776a), gives:
- Clean build, all tests: 10.1s → 5.7s (4.3s in tests)
- Clean build, no tests: 6.0s → 1.7s (72% faster)
- No-change build, no tests: 6.0s → 1.4s (~77% faster)
- Without the `buildtoolset` commands: 0.9s

Various additional minor improvements include:
- `asl` and `asltest` now accept a list of files
- The list of commands in their help output is now sorted
- With -v b8tool prints a startup message showing command and args
  (this helps when a script is running several in sequence)
- Internal documentation improvements
- Minor refactorings