Test: Creating a `.fast-hack` file skips some usually-unused build stuff
This skips several expensive invocations of b8tool that don't usually do
anything when doing "regular" development of stuff assembled by ASL. The
file is never ignored or ocmmitted file so that a `git status` lets you
know that you've not been rebuilding some things that potentially have
changed.

The way to get rid of this hack is to move more stuff into a single
invocation of b8tool, which probably means getting the various
configuration bits it needs out of command-line arguments in ./Test and
into a config file so we can just run b8tool once. (It will still probably
require a small shell script wrapper to deal with setting up/using the
Python virtualenv, however.)