b8tool/bin/b8tool: Remove simple rebuild skip; doesn't work at all
This turns out essentially never to work because the top-level .asm files
we give to to `asl` and the code in `test_rig` in the .pt files we give to
`asltest` both do little more than include another file that has the code
under test. So when we change the file with the code under test, it doesn't
change the file including it and we don't get a rebuild.

Experimentation has shown anyway that, despite that we previously thought
we were saving several seconds, running ASL itself is one of the fastest
parts of the build system, so doing dependency tracking to avoid rebuilds
here offers almost nothing until you're assembling absolutely enormous
amounts of code.