activate → pactivate: Fix bootstrap problem with pip 21.3
Last month's pip upgrade from version 21.2 to 21.3 removed the `-t` option,
which broke the activate script we were previously using. The similar but
new and much better pactivate script, taken from [1], fixes this. It also
now allows us to create a `.python` symlink in the base dir to ensure that
even after a full-clean build we're still using the same Python
interpreter.

`./Test -c` now removes the virtualenv as well as the object files to
provide an easy way of switching to the new interpreter after changing the
`.python` symlink. This unfortunately makes it difficult to do a small-c
clean build without changing the Python interpreter to match `.python`, but
hopefully that won't be a big issue.

[1]: https://github.com/0cjs/pactivate