pactivate → b8tool/pactivate
This should be available for all projects using b8tool since it's a useful
standard way of setting up a Python virtual environment.

However, when it's not in the top level directory we do now need to give it
the project directory explicitly so that it creates/finds `.build/` in the
right place. Another way of approaching this would be to change this
version of pactivate's default directory to $cwd/.., but it's probably
better to keep this exactly the same as the standard version[1] to avoid
confusion and make updates easier.

[1]: https://github.com/0cjs/pactivate