testmc.mos65.machine: Use our IOMem() for memory instead of a list of int
This allows us to use m.setiostreams() and the like to attach functions to
memory locations to simulate memory-mapped I/O, just as we already can do
with testmc.mc6800.Machine.

It's about 15% slower than py65's default list of int, but still between
two and three times as fast as using py65.memory.ObservableMemory.

This changes the underlying store from a list of int to a bytearray, but
py65 seems to be ok with that (at least for all of the hundreds of tests in
this repo that use it), probably because looking up an individual element
in a bytearray returns an int anyway.

Another option we considered was using an ObservableMemory only when we are
asked to add I/O watchers. This might be done transparently by checking to
see if the backing store object responds to subscribe_to_write() and, if
not doing something like `self.memory = ObservableMemory(self.memory)`. But
we feel that added complexity was not worth the relatively minor speed
increase we get when not using memory-mapped I/O in a test, especially
since when we do have it available it we take a big speed penalty on all
memory accesses.