src/pmon/pmon.md: Design for new, rather different pmon monitor
The biggest change from the old one is that we move from reading a full
line of commands and parsing them to a more "immediate" style where we read
no more than a few characters before parsing and (perhaps partially)
executing commands. This not only saves on the memory required for the
command line buffer (giving us hope that this monitor could be usable on a
machine with only 128 bytes of RAM, such as a 6803), but also looks as if
it will make the parsing simpler. In particular, there's much less state to
back out of on errors, so we probably don't need the clever continuation
passing stuff that makes that easier.