Test: Better fast_hack test to see if we should rebuild src/asxxxx/*.a*
Previously when `.fast_hack` was present we always skipped the rebuild of
programs built with asxxxx. This caused an issue with a `-c` rebuild which
(correctly) does not remove .fast_hack but did remove the binaries that
had been generated, and thus caused the tests to fail.

We now use a quick heuristic: if any file we know about seems newer than
its object, we do a full rebuild of all sources using asxxxx. This avoids
builds almost all the time but is fairly likely to do a build if we've
removed the object files or updated a source file. (The heuristic can fail
in serveral ways, but it's good enough for the moment.)