WIP */*/pmon.a68: "Hello, world" version of new pmon for all platforms
pmon itself just prints a banner message and then immediately quits; this
is enough to verify that we're actually loading and running. We also copy
over the `pmon_ramlo` and `pmon_ramhi` definitions from the old pmon, since
dealing with that the same way makes sense. But of course we drop the input
buffer, `rdlinebuf`.

The new tmc68/pmon.a68 header comments now include a handy command line to
run the unit tests and then load it up in the simulator. We should
similarly add to the jr200 and mb6885 files a note on how to generate the
tape, load it and run it, along with whatever tricks one should do to keep
BASIC from stomping on the monitor.