src/mc68/pr: Redefine to assume that `prchar` preserves ABX
This is a reasonable requirement since it's common in home computer BIOSes
that ABX are preserved, and would also be automatic for simple serial port
routines where you can do a bit test (TST) loop and then just write the
output to an address.

Assuming this makes routines such as "read a character, echo it and process
it" considerably more compact and convenient.