WIP src/pmon/: Command letter lookup
This still does not parse any parameters or even wait for a Return
before executing the command.

XXX In fact, we need to fix how `Cerror` works, because when we add
parameter parsing that too will want a Return before showing the command
letter to be an error, whereas we want it to continue to immediately beep
and backspace.