src/mc68/qhex: Optimise away 5 bytes
`qhexword` now uses TST instead of BRA to skip over the `qhexbyte` entry
point to `qhex`, which saves 1 byte but adds 2 cycles.

`qhex` now uses `qhexdigit` instead of `qdigit` which adds 3 bytes if we're
not including `qhexdigit` but saves 4 bytes if we are. (There's currently
no way not to include qhexdigit, and most of the time we're probably
happier with the small chance that we don't use `qhexdigit` elsewhere and
come out ahead overall.