testmc.generic.iomem: Add clearvalue() method to ostreams
This clears the output buffer so that when we run several routines in
a row during a test we need not always match the previous output as
well as any new output.

This may not be the very best way of doing this, but it has the
advantage of working on objects provided by the user if they're
compatible with standard io.IOBase semantics (as BytesIO is).