b8tool.toolset: Extract archive fetcher to setup module
It was previously living as a custom fetch() routine in the asxxxx
module; it's now fetch_archive() and the generic fetch() knows how to
check for the new `source_archive` etc. values on the Setup sublclass.

This currently just fetches the archive and leaves it in
$B8_PROJDIR/.download, as compared to fetch_git() which also
does the "extraction" because that's the way Git works.

This isn't an issue for ASxxxx since it's an archive of binaries that
is extracted by install(), but it's not clear what to do about sources
that we need to extract to a source dir and build.