Test: `b8tool buildtoolset bm2` for the Basic Master Jr. emulator
The workings of this are still somewhat mysterious, but all the source
code is left under .build/tool/src/bm2/, so we can look through it and
figure it out. (There is no documentation, other than what's on the
[bm2 homepage].)

It doesn't seem very amenable to command-line configuration off-hand,
but appears designed to have you typing things at its custom ROM, so
it might be something we want to pull into a GitHub repo and do some
hacking on if we end up needing to automate its use.

[bm2 homepage]: http://ver0.sakura.ne.jp/pc/#bm2