b8tool/bin/mb6885: Run bm2 emulator on a .p file
This currently always loads the MB-6885 ROMs, which come from the osimg
repo installed by the b8tool toolset routines. It would be nice to be able
to start ROM-less with the entry point passed in so we could auto-start a
program that doesn't need the ROMs, but that's going to take a bit more
work.

We might also want to be a bit more sophisticated about having a search
path for bare filenames or "short" paths such as `foo/bar` that could
find, e.g., `{exe,src}/foo/bar.p`.