exe/mb6885/kbscan: Start experiments with scanning keyboard
Right now this just reads the keyboard I/O port in a very tight loop. I
think it's reading all the modifier keys and just one column of the other
keys: Z, A, Q, 7. My guess is that I need to write the row number to scan
to the bottom four bits of the IO_keyboard and re-read to get other rows.
But I need to figure out what the top four bits do on a write.