src/mc68/hello: Also exit after two empty lines in a row, for MB-6885
On the Basic Master Jr. we cannot type control characters (aside from those
sent by the arrow keys), so we've added this alternate method of exiting
the program. (This may be useful on other 6800 systems as well.)

We leave the Ctrl-C method of exiting in place for no particularly good
reason.