b8tool/bin/mb6885: -n option to run without loading ROMs
BUG: Start vector is hard-coded to $3000.

The bm2 emulator can emulate a minimal set of the ROM subroutines if the
ROM images are not provided. It can also in this mode load the reset vector
to "autostart" your program, which it cannot do if you're using ROM images.

We allow the `-n` option to run without ROM images and thus supply a start
vector, which makes manual testing faster. However, though we should be
taking the start vector from the S-record file, we don't do that yet, so
this always runs whatever's at $3000 when run with an S-record file in
ROMless mode.