exe/mb6885/hello: Enable BREAK key and update documentation
The BREAK key didn't work when this was called from the monitor because the
monitor seems to disable at at some point before running this, though we're
not quite sure where. Re-enable it with NMICLR and explain its still funny
behaviour.

Also update the documentation for calling from BASIC, the new 2× Enter
method of exiting and a few other things.