WIP cmtconv..pc8001: Block objects and BASIC blocks read mostly done
We now have Block objects for BASIC header block, BASIC text block and
binary save block.

We read the minimum length of any initial block which can tell us whether
we're reading a BASIC file or monitor file. We attempt to create a BASIC
header block and, if that works, we continue on to read the BASIC text
block. We read what we can, but have not quite figured out what's going
on with the trailing material at the end.

If we fail to create a BASIC header block, we have a binary save block,
for which we do nothing at the moment but raise a "Write me!"