src/i8080/seg7test: Fix broken program
The program never terminated because I was shifting with _copy_ to the
carry rather than _through_ the carry, so there was always a '1' bit in A.
Thus if left running it eventually overwrite all of memory, including
itself, and then crashed.

Switch to rotating _through_ the carry so that we can start with all zeros
and get back to all zeros. We preload the carry with 1 before starting so
that we have a bit come in, and the later `or a` test not only updates the
zero flag for our branch test, but also clears the carry to ensure that we
never shift another '1' bit in.