src/qdigit.py: Extract in/out register and ok/error flag to a fixture
The i8080 does not have an N flag, so we'll need to use a different
one for the ok/error indication of the qdigit routine. To handle this,
we extract a fixture giving a function that takes the in/out register
value and True/False for ok/error and returns a Machine.Register object
with the machine-specific register and flag set appropriately.