tmon.i80: command parser, Q command, start on E command
This includes src/i8080/tmon.i80 with the machine-independent code and
exe/tmc/i8080/tmon.i80 to run it on the command-line simulator.

This handles all command read, error checking, execute and cancel, and
the Q (quit) command, which has no options, that simply returns if
executed. There's also a dummy E (examine memory) command that will
simply print an 'E'.

Any attempt to enter an option for a command will halt the machine.
(The simulator throws an exception as the `halt` command isn't
implemented.)