i8080/pr: prspace, prvischar and prhex routines
The code is a near direct translation of the mc68/pr.a68, and the
tests are virtual duplicates if its tests. We should figure out how to
refactor out the duplication at some point. In the mean time, there
are also a few minor changes to mc68/pr.a68 to make the duplication
more clear.

Note that there's an API diference: prvischar here preserves A (as it
should do for the other CPUs) because it's very typical that the
caller wants to do something with the character after echoing it,
and prvischar doesn't necessarily pass the same value to prchar
as it receives.