tool/t8dev: Update to v0.2.0 (994a671)
Update tests that relied on call() not executing return instructions. There
are various spots where we were trying to confirm which return instruction
was executed that we now need to check via stopping at that specific
location, rather than relying on call() to stop before executing the opcode
at that location.

Change bios.* includes to biosdef.*.

Add r8format, fixed to version 0.2.x and t8dev to requirements.txt.
Note that we cannot use ~= 0.2 here because Pip does not follow semantic
versioning for 0.x releases, it considers 0.3 to be compatible with 0.2.