i8080/prvischar/test.py: Extract common test function
This is not quite right in that we still end up defining a
test_prvischar() in every .pt file that does nothing but call the
common prvischar_test(), but it's not clear at this time how to do
this more cleanly. Ideally test.py when imported could just pick up
the CHAR_OUTPUT for the particular module in which it's imported, but
decorators don't really work that way; there's probably some other way
in the pytest API to generate the tests.