src/generic.py,i8080/prscreenchar: Screen char printing
The src.i8080.prvischar.test.prvischar_test() routine turns out to be
a generic "call routine with value in A and check printed output"
routine, so we move it to src.generic.a_output_test().

Screen char printing is essentially a placeholder at the moment,
as we provide only "ASCII" output.