../tmon/README: Rewrite for more realistic input methods and new commands
Having started on an implementation of this monitor, it's clear that the
input can be done with less memory if we change the way a few things work,
including abandoning the idea of using rubout (backspace or DEL) to delete
a previous character.

"Option" was kind of the wrong word for the command parameters, because
mostly they're still paramters that are used for commands even if you don't
specify a new value when typing in the command, so we go back to saying
"parameter" and "parameter value."

We've swapped `<` and `>`, using the former for byte entry and the latter
for word entry, because obviousl the one saying "smaller" should be the
byte entry and the one saying "greater" should be the word entry. These
also match the convention used for LSB and MSB in 6502 assembler.

This adds a summary of commands in alphabetical order, which makes it much
easier to determine what commands are used and free, adds /, f, v, c, o, p,
l and w, changes a to =, and adds a symbol to show which parameters are
memorised (which is most of them, sadly).

The Intel hex and Motorola S-record formats are essentially deposit
commands that always have an address parameter, so we move them up to just
after the regular deposit command. (The "common updating address" commands
`<`, `>` and `'` are after all other deposit commands.)

The "Load/Save" section becomes "Load from and Write to Devices" and we add
the `l` (load) and `w` (write) commands. You'd think we'd use l/s for
load/save, or r/w for read/write, s conflicts with Motorola S-record
format, and r conflicts with the registers command.