tmon/README: Many further updates and reworking of deposit UI
The main change is that we no longer have separate commands for "keep start
address" and "advance start address"; these are simply modes of the byte
(renamed to `.`), word (`>`) and ASCII (`'`) deposit commands. Thus we no
longer need a `d` command for deposit, and that is repurposed to set
deposit options for the other three commands.