i8080/tmon: Command labels now named for function rather than keystroke
We're looking soon at changing some of the command letters, which
would invalidate labels based on command letters, so instead use
labels based on command functionality so that they need not change
if/when we change the letters.

This also introduces a `cmd` macro to generate the boilerplate code
to dispatch a command character to its function.