WIP i8080/tmon: Switch to command table
This will let us put the address of the parameter table in the command
table, which means we no longer need to load HL with it at the start of
every command. And maybe not have to `call read_params`, either.

But this only makes the code smaller if we use less space for the routine
that reads and executes this table than the HL loads (11 commands × 1 byte)
and the calls (11 commands × 4 bytes). Obviously it's not going to help for
param table loads alone (11 bytes), but surely we can do the whole thing in
well under 55 bytes, if the call to read_params can also be extracted.

Unfortunately we have some commands that need to set up stuff _before_
read_params is called:
- cmd_iret, cmd_jump, cmd_call: all need a setup and push on stack
- cmd_deposit_params: needs to clear a temp location before params