Test: Add INTERNAL ERR, FAILED and OK notices, and time the run
We add a `trap … ERR` that displays the failing line number, command,
etc. when the script unexpectedly exits. This now requires that we
have a fail() function and add `|| fail` after commands like pytest
that we expect sometimes to fail (as opposed to it being an error in
our script) so that it's not marked as an internal error.

fail() prints FAILED and gives the error code, and we now print OK
when nothing fails. In both cases we give the elapsed time of the
entire run as well, to 1-second granularity.