Test: -E option to run an emulator after pytest passes
This introduces the post_emulate_run() function; that and the -E parsing
can probably be moved to t8setup.bash once we've played around with this
for a while and the UI and API have settled down.

Currently this knows how to run only tmc (command line testmc simulator),
and it extracts that we should use tmc and the CPU to use from the path
(.build/obj/exe/tmc/CPU/…). But adding `t8dev emulate` support shouldn't be
too bad once we figure out how we extract that information too from the
path, perhaps changing our paths to match this sort of thing.

Also update to latest t8dev, which now leaves the first -- option for us
to see, rather than eating it. We don't actually need this, though.