i8080/std.i80: Mostly-done standard setup, and everything uses it now
The test for the `db ...` is commented out, however because while the
`db` syntax is native the tests use the Motorola repetition syntax
`[n]X` because even with our macros Motorola doesn't accept `n dup X`.

So looks like we need to create our own function to do the "dup" thing.