exe/cpm/tmon100: First go at porting tmon to CP/M: BROKEN
The CPM BDOS functions are rather unlike the "standard" BIOS functions we
use in that they need a code loaded into C to indicate which function is
being requested, and they don't preserve any registers (as far as I can
tell). So we need to write our own "mini-BIOS" wrappers around these.

Unfortunately, at startup tmon prints a `!` ($21) instead of a `.` ($2E),
and things get worse from there, with what are supposed to be single
characters printed as a different character multiple times. This needs to
be debugged.