exe/cpm/tmon100: Move BIOS read/print functions to testmc/i8080/bios/cpmcode
That version in t8dev has a bunch of fixes, too, but it still doesn't
work quite right for CP/M tmon; there are some backspace issues that
don't appear in the tmc version. It's not clear whether that's really
the CP/M version of the generic BIOS, since some basic tests of
backspace seem to work fine.