t8dev: 0.3.0.dev7 (cabd10e) and fix tmon prBchars bug this surfaced
This version of t8dev clarifies the generic monitor API to indicate
that `prchar` may destroy A, and changes the implementation actually
to do that (even though it's not necessary) to help surface bugs.

In tmon, `prBchars` assumed that A was preserved and here now that
it's not, the error in it shows up in tests. It was easily fixed.