exe/cpm/tmonlvt → tmonlo: Use configurable prvischar
We no longer need to mark this a VT100-specific because, though we still
default to VT100 attribute codes (because that's the most common serial
terminal in these days of Unix machines running xterm), they can now
be easily changed by patching the first few bytes of the file.

This can even be done on the fly in the running tmon.