src/stdtest*: Remove `dup` functionality from dbtest
The "duplicate this byte N times" syntax is different for Motorola
`byt` and Intel `db` operands, so even though we can redefine `byt` to
`db`, the syntax is still not entirely compatible. We're working
on ways to deal with thus, but so far a straight function substitution
doesn't work because that works on values, not at the text level.

In the meantime, we just remove the test for `[3]$FF` because we
can't do an equivalent `3 dup $FF` in the same statement, and just
use `$FF, $FF, $FF` instead, while leaving in comments explaining
the problem.