i8080/tmon.pt: Add tests for byte deposit mode, exposing bug
It turns out that in byte deposit mode with advance-to-next turned on,
starting a deposit and entering just CR still moves the deposit start
forward. For the moment we've xfailed this test.

This involved some refactoring of test_deposit_ascii() to allow it to test
byte and word deposit modes; thus it's renamed to test_deposit(), and the
other test_deposit_*() functions that were actually testing the settings
are now test_depset_*().