i8080/tmon: print regs o nentry; `intentry` prints ! as well
We generally want to see the registers on entering the monitor, so always
print them. In most cases, when testing a routine, you want to see the
registers after it's been executed anyway, so this avoids having to type
the `R` every time. (It does use an extra line on the screen; perhaps we
can think of some way to avoid that in the--probably rare--cases where it's
annoying rather than helpful.)

The `!` in front of the register on `intentry`  makes it clear that we
entered via an interrupt (or the RST that is usually set up as a breakpoint
marker).

This has exposed some bugs: though the call command sets up `entry` as the
return address, we're seeing the `!` anyway when calling, or even jumping
to, a RET ($C9). We appear to be placing $0000 on the stack as the return
address, which would be executing the NOPs in uninitialised memory until we
hit the `JMP intentry` at RST $08. Though oddly it doesn't seem to be
updating the saved user PC....