*: Update to ASL build 280, which fixes (mostly) `db` duplication syntax issue
This has exposed a bug in `dw` for MC6800 where it uses little-endian
rather than native-endian order. We work around this for the moment
by keeping our `dw` macro for that CPU, but we still can get rid
of all the other `ds/db/dw` macros.

Also add some notes to DEVEL.md about that bug and that we should
be checking assembler version.