Test,exe/nec/pc8001/hello: Run CSCP NEC emulator with expansion ROM
If the -E parameter is exe/nec/, we run a CSCP emulator with the name
taken from the next component down in the directory structure.
Currently we hard-code this to patching N80.ROM at $6000, which is
the expansion ROM area on the PC-8001; we'll have to expand the
processing of this for other machines.

We also need to look into how we might use other data setup methods,
such as setting up CMT or disk images, and perhaps even memory images.

The t8dev/ update is a fix for the error exits of t8dev emu cscp.