exe/kc85/m100/hi.i85: Test of building/running TRS-80 Model 100 .CO programs
This updates Test to run the virtualt emulator with the -E option as well.
However, our way of building the .CO files is, at best, a terrible hack.
What we really need is to replace ASL's `p2bin` with a new `p2b` program in
`r8format` that can build arbitrary "binary program file" headers, and has
a set of platform options (`-p kc85` etc.) that will do the right thing,
including a default output filename with the proper extension.