exe/kc85/m100/tmon: Initial port of tmon to TRS-80 Model 100
TODO:

We're currently using straight-ASCII prvischar. This should be changed
to use inverse (via ESC-p/ENTREV and ESC-q/EXTREV) for control
characters (including $7F) and the straight value for $80-$FF, which
all print as special graphic characters. It appears that there are no
glyphs for codes $00-$1F (as there's no actual frame buffer in the
machine--the ouput routines generate bitmaps to the display's separate
frame buffer), so prscreenchar can just be the same as prvischar.

The XXX notes for errbeep and exit should be looked at to see about
changes to bring back into the system-independent tmon code.

And of course this still requires a user CLEAR (not sure that there's
any way around that) and uses 1.5-2K of space at the top of memory
when it would be better to run it from "inside" file, after tweaking
it for reloc.