exe//pc8001/exprom1: Entirely rework I/O and various other things
After further investigation and disassembly of the PC-8001 system ROM,
we now distinguish between printing to the screen and printing to the
current output device (which defaults to the screen); we now use just
the former character output routine. We've renamed things appropriately.

We now init all the Disk BASIC command hooks to our own routine that
just prints the registers so we can see what commands lead us to
what hooks, and then override just `CMD` to do our simple "hi."

Add a check to see we've not overflowed the ROM, and comments
indicating how we can switch MON to a different monitor program.
(Though we will probably leave that alone and instead use `CMD TMON`
to enter tmon when we add that.)