i8080/tmon: Int entry char now `^`; add regular entry char `.`
This is the character that preceeds the register dump done on entry.
Originally, it was just to be an indication that we'd entered the monitor
via the interrupt entry point (and thus have a PC to continue at), but it
turns out to be useful to help separate any program output from the
monitor's register dump, too.

So we give the regular entry point a preceding `.`, and change the
interrupt entry point char from `!` to `^`, because the latter is more
easily distinguished from the numbers to the right of it (and possibly to
the left as well).

This now also solves the TODO issue about distinguishing program output
from monitor output on entry.