i8080/tmon cmd_memcopy: `m` command implementation
This implements only forward copies at the moment, and so will work for
overlapped ranges only if the destination is earlier in memory than the
source. This needs to be fixed.

There are also some "utility" math routines that probably need to be
generalised and/or extracted, but that's left for after we do the overlap
fix since that will change some of these. (And we also need to look for
other code that is or should be using these. Possibly the 16-bit math
routines for the `/` (calculate) command?)

It also doesn't handle certain overlapping situations where the source is
shorter than the destination and thus looped, but there's no good way to
handle this because we'd need to use "external" storage somewhere for a
copy of the source. So for that we just need to add a warning to the
documentation.