exe//pc8001/exprom1: Tweak tmon_ram and document issues with it
We lower tmon_ram because our work area was definitely going way
past $EA00 and into the ROM work area with the old values; now
it goes up to just $E9FB. But actually we need a better check that
this is not overflowing.

Even after a CLEAR 100,58000, though, our saved values are still
being destroyed; we need to figure out why this is.