exe//pc8001/exprom1 tmon: Use top memory area for tmon_ram
At reset/RST 00 the BIOS sets the stack to $FFFF, but that's maintained
only through the end of expansion ROM init; after that it's moved down
into the BIOS/BASIC work area. The internal MON command uses memory up
to $FF3C, which means that (as far as we can tell from disassembling
the code), $FF3D through $FFFF are entirely unused while the system is
running.

Thus we start our tmon work area at $FF3D and have the saved regs,
variables, etc. starting at $FF80. We allocate this much stack (67
bytes) because the BIOS routines we call seem to use a lot of stack;
we've seen tmon use 42 bytes when calling prchar.

This leaves tmon_ram_end currently at $FFAB, giving 85 free bytes
above that; we choose to leave these at the top of memory rather than
the bottom on the assumption that machine-langauge programmers who are
going to choose a random bit of memory to store a few words will
probably use the very top of RAM.