WIP i8080/tmon,exe//pc8001/exprom1: Add init_platform for PC-8001
This requires at least ASL bld 283 for the symexist() function.

The optional `init_platform` routine, if it exists, is called by tmon
just after initialising the work area, allowing additional code to do
further platform- and situation-specific setup. The unit tests test
the case where `init_platform` is present, but not absent. (That's
been tested by hand, and also is tested in the exe/… versions that
don't use `init_platform`.)

The PC-8001 version uses this to set the examine width to 8 (instead
of the default 16) if the system is in 40-column mode, which stops the
annoying wrapping of examined data.

XXX implement the PC-8001 thing above