i8080/tmon,exe//pc8001/exprom1: Add init_platform, PC-8001 default width fix
This requires at least ASL bld 283 for the symexist() function.

The optional `init_platform` routine, if it exists, is called by tmon
just after initialising the work area, allowing additional code to do
further platform- and situation-specific setup. The unit tests test
the case where `init_platform` is present, but not absent. (That's
been tested by hand, and also is tested in the exe/… versions that
don't use `init_platform`.)

The PC-8001 version uses this to set the examine width to 8 (instead
of the default 16) if the BIOS current screen width variable
(`crt_width`) is less than 72, annoying wrapping of examined data.

(This actually doesn't quite work in 72-column mode because the
16-width control code at the end of the line, which should bring the
cursor back to dump, with ASCII, is exactly 72 columns wide; the
cursor then wraps to the line, producing a blank line between each
line of memory display. We can move the cursor to the next line
without affecting the character previously next line after the last
character and the newline moves down one further probably handle this
by having the examine command print a single move left the end of the
previous line, and then printing the newline, which will under the
cursor. And that will handle both the default 72 line case and any
cases where, e.g., the particular dump configuration gives a line of
exactly 40 or 80 characters.)

This includes minor cleanups of PC-8001 BIOS equates and the PC-8001
tmon port to-do list comment.